requirements.txt
opencv-python
numpy
RPi.GPIO
mfrc522
flask
flask-cors
pyserial
